Add DataSig verifier
