+++ src/utils/data-sig/data-sig.js
@@ -1,4 +1,5 @@
 const { signer } = require('../../lnd-rpc/signer')
+const { getMyAddress } = require('../../utils/generic')
 var protobuf = require("protobufjs")
 
 const generateDataSig = (version, data, destinationAddress, senderAddress) => {
@@ -81,9 +82,32 @@ const decodeDataSig = (dataSigBuffer) => {
     })
 }
 
+const verifyDataSig = (dataSigBuf, dataBuf) => {
+    return new Promise(async function (resolve, reject) {
+        const dataSig = await decodeDataSig(dataSigBuf)
+
+        const myAddr = await getMyAddress()
+        const myAddrBuf = Buffer.from(myAddr, 'hex')
+
+        const destDataBuf = Buffer.concat([myAddrBuf, dataBuf])
+
+        let request = {
+            msg: destDataBuf,
+            signature: dataSig.sig,
+            pubkey: dataSig.senderPK
+        };
+
+        signer.verifyMessage(request, function (err, response) {
+            if(err) reject(err)
+            if(response) resolve(response.valid)
+        });
+    })
+}
+
 
 module.exports = {
     encodeDataSig,
     decodeDataSig,
-    generateDataSig
+    generateDataSig,
+    verifyDataSig
 }
